==> course/08_named_entity_recognition/01_getting_reddit_data.ipynb <==
# Getting Reddit Data

There are two options for extracting data from Reddit:

* The `requests` library, which will allow us to interface directly with the Reddit API.

* The PRAW library, which is a wrapper library that adds an extra layer of abstraction in accessing the Reddit API.

Here we will cover the first option, using the `requests` library to interface directly with the API.

The final extraction script will look like this:

import requests

def reading(file_name = 'credentials.txt'):
    s = open(file_name, 'r').read()
    dict = eval(s)
    return(dict)

dict = reading()

#### Commented code is from the lecture, but results in `'invalid_grant'` error

auth = requests.auth.HTTPBasicAuth(dict['client_id'], dict['secret'])

data= {'grant_type': 'password',
        'username': dict['user'],
        'password': dict['pwd']}

headers = {'User-Agent': 'MyBot/0.0.1'}

res = requests.post(f'https://www.reddit.com/api/v1/access_token',
                            auth=auth, data=data, headers=headers)

res

# Output: <Response [200]>

res.json()

# Output: {'error': 'invalid_grant'}

token = res.json()['access_token']

headers['Authorization'] = f'bearer {token}'

headers

requests.get('https://oauth.reddit.com/api/v1/me', headers=headers)

api = 'https://oauth.reddit.com'

res = requests.get(f'{api}/r/hawwkey/new', headers=headers)

# Commented to reduce jupyter notebook size
# res.json()

res = requests.get(f'{api}/r/hawwkey/new', headers=headers, params={'limit': '100'})

# Commented to reduce jupyter notebook size
# res.json()

name, created_utc, subreddit, title, selftext, upvote_ratio, ups, downs, score

import pandas as pd

df = pd.DataFrame({
    'name': [], 
    'created_utc': [], 
    'subreddit': [], 
    'title': [], 
    'selftext': [], 
    'upvote_ratio': [], 
    'ups': [], 
    'downs': [], 
    'score': []
})

for posts in res.json()['data']['children']:
    df = df.append({
        'name': posts['data']['name'], 
        'created_utc': posts['data']['created_utc'], 
        'subreddit': posts['data']['subreddit'], 
        'title': posts['data']['title'], 
        'selftext': posts['data']['selftext'], 
        'upvote_ratio': posts['data']['upvote_ratio'], 
        'ups': posts['data']['ups'], 
        'downs': posts['data']['downs'], 
        'score': posts['data']['score']
}, ignore_index=True)

df.head()

df['name'].iloc[len(df)-1]

while True:
    res = requests.get(f'{api}/r/hawwkey/new', headers=headers, 
                       params={'limit': '100',
                               'after': df['name'].iloc[len(df)-1]})
    if (len(res.json()['data']['children'])==0):
        break
    for posts in res.json()['data']['children']:
        df = df.append({
            'name': posts['data']['name'], 
            'created_utc': posts['data']['created_utc'], 
            'subreddit': posts['data']['subreddit'], 
            'title': posts['data']['title'], 
            'selftext': posts['data']['selftext'], 
            'upvote_ratio': posts['data']['upvote_ratio'], 
            'ups': posts['data']['ups'], 
            'downs': posts['data']['downs'], 
            'score': posts['data']['score']
        }, ignore_index=True)

df.tail()

df = df.replace({'|': ''}, regex=True)

df.to_csv('./data/reddit_hawwkey.csv', sep='|', index=False)

res.json()['data']['children'][0]['data']

res.json()['data']['children'][0]['data']

#### Using the following class, however, works just fine:

import requests
import pandas as pd


class Reddit:
    def __init__(self, client_id, secret_token, username, password):
        # first create authentication object
        auth = requests.auth.HTTPBasicAuth(client_id, secret_token)
        # build login dictionary
        login = {'grant_type': 'password',
                 'username': username,
                 'password': password}
        # setup header info (incl description of API)
        headers = {'User-Agent': 'MyBot/0.0.1'}
        # send request for OAuth token
        res = requests.post(f'https://www.reddit.com/api/v1/access_token',
                            auth=auth, data=login, headers=headers)
        # pull auth bearer token from response
        token = res.json()['access_token']
        # add authorization to headers dictionary
        headers['Authorization'] = f'bearer {token}'
        # add headers dict to internal attributes
        self.headers = headers
        # and api
        self.api = 'https://oauth.reddit.com'

    def get_new(self, subreddit, iters):
        # initialize dataframe to store data
        df = pd.DataFrame()
        # initialize parameters dictionary
        params = {'limit': 100}
        # iterate through several times to make sure we get all the data available
        for i in range(iters):
            # make request
            res = requests.get(f'{self.api}/r/{subreddit}/new',
                               headers=self.headers,
                               params=params)
            # check that we returned something (if not we reached end)
            if len(res.json()['data']['children']) == 0:
                print('No more found')
                return df
            # iterate through each thread recieved
            for thread in res.json()['data']['children']:
                # add info to dataframe
                df = df.append({
                    'id': thread['data']['name'],
                    'created_utc': int(thread['data']['created_utc']),
                    'subreddit': thread['data']['subreddit'],
                    'title': thread['data']['title'],
                    'selftext': thread['data']['selftext'],
                    'upvote_ratio': thread['data']['upvote_ratio'],
                    'ups': thread['data']['ups'],
                    'downs': thread['data']['downs'],
                    'score': thread['data']['score']
                }, ignore_index=True)
            # get earliest ID
            earliest = df['id'].iloc[len(df)-1]
            # add earliest ID to params
            params['after'] = earliest
        return df

SUB = 'hawwkey'

CLIENT_ID = dict['client_id']
SECRET_TOKEN = dict['secret']

USER = dict['user']
PWD = dict['pwd']

reddit = Reddit(CLIENT_ID, SECRET_TOKEN, USER, PWD)

data = reddit.get_new(SUB, 20)

data = data.replace({'|': ''}, regex=True)

data.to_csv(f'./data/reddit_{SUB}.csv', sep='|', index=False)

data.head()
