==> course/08_named_entity_recognition/01b_ner_reddit_freq_table.ipynb <==
# NER on Reddit Data

In this example we work through building a NER tool for extracting organization mentions from Reddit data to create a frequency table.

import spacy
from spacy import displacy
import pandas as pd
from collections import Counter

nlp = spacy.load('en_core_web_sm')

The `get_orgs` function will consume a length of text, and return a list of mentioned organizations:

df = pd.read_csv('./data/reddit_hawwkey.csv', sep='|')
df.head()

def get_orgs(text):
    # process the text with our SpaCy model to get named entities
    doc = nlp(text)
    # initialize list to store identified organizations
    org_list = []
    for entity in doc.ents:
        # here we modify the original code to check that entity text is not equal to one of our 'blacklisted' organizations
        # (we also add .lower() to lowercase the text, this allows us to match both 'nyse' and 'NYSE' with just 'nyse')
        if entity.label_ == 'ORG':
            org_list.append(entity.text)
    # if organization is identified more than once it will appear multiple times in list
    # we use set() to remove duplicates then convert back to list
    org_list = list(set(org_list))
    return org_list

df['organizations'] = df['title'].apply(get_orgs)

df.head()

from collections import Counter

orgs = df['organizations'].to_list()

# Convert list of lists to a single list
orgs = [org for sublist in orgs for org in sublist]

# Commented to reduce file size of notebook
# orgs

org_freq = Counter(orgs)

org_freq.most_common(10)

BLACKLIST = ['nhl', 'instagram', 'gm', 'tsn']

def get_orgs_blacklist(text):
    # process the text with our SpaCy model to get named entities
    doc = nlp(text)
    # initialize list to store identified organizations
    org_list = []
    for entity in doc.ents:
        # here we modify the original code to check that entity text is not equal to one of our 'blacklisted' organizations
        # (we also add .lower() to lowercase the text, this allows us to match both 'nyse' and 'NYSE' with just 'nyse')
        if entity.label_ == 'ORG' and entity.text.lower() not in BLACKLIST:
            org_list.append(entity.text)
    # if organization is identified more than once it will appear multiple times in list
    # we use set() to remove duplicates then convert back to list
    org_list = list(set(org_list))
    return org_list

df['organizations'] = df['title'].apply(get_orgs_blacklist)

orgs = df['organizations'].to_list()
orgs = [org for sublist in orgs for org in sublist]
org_freq = Counter(orgs)
org_freq.most_common(10)

df['organizations'].head()

df.to_csv('./data/reddit_hawwkey_ner.csv', sep='|', index=False)

## Applying NER

All we need to do now is load in the */r/investing* data and apply the `get_orgs` function to our text column to create a new `organizations` column.

1. Load the data and view the top five rows with `df.head()`:

df = pd.read_csv('./data/reddit_investing.csv', sep='|')
df.head()

2. Extract mentioned organizations from `selftext` and add to a new column called `organizations`:

df['organizations'] = df['selftext'].apply(get_orgs)
df.head()

*(This step can take a long time to run. It can be useful to break larger datasets into more manageable chunks if required)*

Now we have a list of all mentioned organizations contained within the `organizations` column. We will now take the full column, merge each list, and use `Counter` to create a frequency table of organization mentions.

# merge organizations column into one big list
orgs = df['organizations'].to_list()
orgs = [org for sublist in orgs for org in sublist]
orgs[:10]

## Create Frequency Table

from collections import Counter

# create dictionary of organization mention frequency
org_freq = Counter(orgs)

We now have a *Counter* dictionary containing all of our organization labels as *keys*, and their mention frequency as *values*. The `most_common(n)` method allows us to view the **n** most frequently mentioned organizations:

org_freq.most_common(10)
