==> course/08_named_entity_recognition/03_ner_and_sentiment.ipynb <==
# NER and Sentiment

In this section we will work through applying basic sentiment analysis to our data using a pre-built *distilBERT* model from the **Flair** library. We will then use our organization labels captured through NER in the previous section to create a list of organizations with the highest and lowest average sentiment scores.

import pandas as pd
import flair

We initialize the English sentiment model `en-sentiment`:

model = flair.models.TextClassifier.load('en-sentiment')

For each sample there are a few steps we need to take to create the sentiment score. We need to tokenize the input text, make a prediction, extract the direction (*positive* or *negative*) and confidence (a score from 0 to 1). If this is new to you, we cover the Flair sentiment model in more depth in **TK insert link**.

The following function carries out each of these steps for a single extract:

def get_sentiment(text):
    # tokenize input text
    sentence = flair.data.Sentence(text)
    # make sentiment prediction
    model.predict(sentence)
    # extract sentiment direction and confidence (label and score) object
    value, score = sentence.labels[0].value, sentence.labels[0].score
    return [value, score]

We now need to load our previously processed dataframe (which includes the *organizations* column) and `apply` the `get_sentiment` function to the *selftext* column. These sentiment scores will then be stored in a new *sentiment* column.

# load data
df = pd.read_csv('./data/reddit_hawwkey_ner.csv', sep='|')
df.head()

# get sentiment
df['sentiment'] = df['title'].apply(get_sentiment)

df['sent_value'] = df['sentiment'].apply(lambda sentiment: sentiment[0])
df['sent_score'] = df['sentiment'].apply(lambda sentiment: sentiment[1])
df.head()


type(val_sco_list)

df['sentiment'][0].value

Now we need to extract each of the organizations alongside it's sentiment score. We will then loop through each, tallying up a total sentiment score and count.

Before we do that, we need to convert each value in the *organizations* column to a list (they are currently strings because we cannot save Python lists to file within Pandas dataframes, they are automatically converted to strings).

import ast

df['organizations'] = df['organizations'].apply(lambda x: ast.literal_eval(x))

# initialize sentiment dictionary
sentiment = {}

# loop through dataframe and extract org labels and sentiment scores into sentiment dictionary
for i, row in df.iterrows():
    # extract sentiment direction and score
    direction = row['sent_value']
    score = row['sent_score']
    # loop through each label in organizations column
    for org in row['organizations']:
        # check if org label exists in sentiment dictionary already
        if org not in sentiment.keys():
            # if it doesn't, initialize new entry in dictionary
            sentiment[org] = {'POSITIVE': [], 'NEGATIVE': []}
        # append positive/negative score to respective dictionary entry
        sentiment[org][direction].append(score)

df['organizations']

sentiment['Rangers']

Now we can loop through each organization entry in the sentiment dictionary and calculate an average positive, and average negative score:

# initialize sentiment list
avg_sentiment = []

for org in sentiment.keys():
    pos_freq = len(sentiment[org]['POSITIVE'])
    neg_freq = len(sentiment[org]['NEGATIVE'])
    for direction in ['POSITIVE', 'NEGATIVE']:
        score = sentiment[org][direction]
        if len(score) == 0:
            sentiment[org][direction] = 0.0
        else:
            # otherwise calculate total
            sentiment[org][direction] = sum(score)
    total = sentiment[org]['POSITIVE'] - sentiment[org]['NEGATIVE']
    avg = total/freq
    pos_avg = sentiment[org]['POSITIVE']/pos_freq if pos_freq !=0 else 0
    neg_avg = sentiment[org]['POSITIVE']/neg_freq if neg_freq !=0 else 0
    avg_sentiment.append({
        'entity': org,
        'positive': pos_avg,
        'negative': neg_avg,
        'frequency': pos_freq+neg_freq,
        'score': avg
    })

avg_sentiment[:3]

sentiment_df = pd.DataFrame(avg_sentiment)
sentiment_df.head()

sentiment_df = sentiment_df[sentiment_df['frequency'] > 3]
sentiment_df.head()

sentiment_df.sort_values('score', ascending=False).head(10)

sentiment_df.sort_values('score').head(10)

# initialize sentiment list
avg_sentiment = []

# loop through each organization
for org in sentiment.keys():
    # get number of positive and negative ratings
    freq = len(sentiment[org]['POSITIVE']) + len(sentiment[org]['NEGATIVE'])
    for direction in ['POSITIVE', 'NEGATIVE']:
        # assign to variable for cleaner code
        score = sentiment[org][direction]
        # if there are no entries, set to 0
        if len(score) == 0:
            sentiment[org][direction] = 0.0
        else:
            # otherwise calculate total
            sentiment[org][direction] = sum(score)
    # now calculate total amount
    total = sentiment[org]['POSITIVE'] - sentiment[org]['NEGATIVE']
    # and the average score
    avg = total/freq
    # add to sentiment list
    avg_sentiment.append({
        'entity': org,
        'positive': sentiment[org]['POSITIVE'],
        'negative': sentiment[org]['NEGATIVE'],
        'frequency': freq,
        'score': avg
    })

sentiment_df = pd.DataFrame(avg_sentiment)
sentiment_df.sort_values('frequency', ascending=False).head(10)

Immediately we can see we have a lot of entities which have appeared once in our dataset, and because of this their score will be pushed to one extreme or the other. We can filter out anything with less than or equal to a frequency of `3` to remove many of these instances:

sentiment_df = sentiment_df[sentiment_df['frequency'] > 3]
sentiment_df

Here we have some more relevant information. We can see a few items that we can remove through the `BLACKLIST` covered in earlier sections such as *Fed* and *Treasury*, but nonetheless this list is looking much better than before. We can apply `sort` to search for the entities with the highest overall score:

sentiment_df.sort_values('score', ascending=False).head(10)

Very quickly we've got our results that we have pulled together using simple, ready-to-use models, and **zero** text preprocessing. With further fine-tuning, and process development, these already good results can become great. Which we will cover soon.
