==> course/08_named_entity_recognition/04_spacy_and_transformers.ipynb <==
# NER And Transformers

import spacy
from spacy import displacy

!nvcc --version

!pip install spacy[transformers,cuda117]

!python -m spacy download en_core_web_trf

nlp = spacy.load('en_core_web_sm')

trf = spacy.load('en_core_web_trf')

doc = nlp("Apple reached an all-time high stock price of 143 dollars this January.")

displacy.render(doc, style='ent')

---

doc = trf("Apple reached an all-time high stock price of 143 dollars this January.")

displacy.render(doc, style='ent')

txt = "Total nonfarm payroll employment rose by 266,000 in April, and the unemployment rate was little changed at 6.1 percent, the U.S. Bureau of Labor Statistics reported today. Notable job gains in leisure and hospitality, other services, and local government education were partially offset by employment declines in temporary help services and in couriers and messengers."

doc = trf(txt)
displacy.render(doc, style='ent')

doc = nlp(txt)
displacy.render(doc, style='ent')

txt = """Fastly released its Q1-21 performance on Thursday, after which the stock price dropped a whopping 27%. The company generated revenues of $84.9 million (35% YoY) vs. $85.1 million market consensus. Net loss per share was $0.12 vs. an expected $0.11.

These are not big misses but make the company one of the few high-growth cloud players that underperformed market expectations.

However, the company also lowered its guidance for Q2: Fastly forecasts revenues of $84 - $87 million and a net loss of $0.16 - $0.19 per share, compared to the market consensus of $92 million in revenue and a net loss of $0.08 per share, thereby disappointing investors.

Lastly, Adriel Lares will step down as CFO of the company after 5 years."""

doc = trf(txt)
displacy.render(doc, style='ent')

doc = nlp(txt)
displacy.render(doc, style='ent')
